Migrate dfu/test-device-fw-compatibility to pytest
